video_rag/requirements.txt

torch
transformers
accelerate
openai-whisper
easyocr
ultralytics
sentence-transformers
faiss-cpu
opencv-python
numpy
pillow
